// src/app/shared/models/persona-juridica-interface.ts
export interface PersonaJuridicaInterface {
    tipoCliente: string;
    tipoPersona: string;
    tipoIdentificacion: string;
    razonSocial: string;
    identificacion: string;
    nombreContacto: string;
    cargoContacto: string;
    anioInicioActividad: number;
    email: string;
    numeroCelular: string;
    nacionalidad: string;
    clave: string;
    tipoContacto: string;
    aceptaPoliticaPrivacidad: string;
    aceptaTerminoUso: string;
}